가게에 리뷰 추가

+++ src/main/java/umc/spring/web/dto/review/ReviewRequestDTO.java
@@ -12,7 +12,7 @@ public class ReviewRequestDTO {
     @Getter
     public static class JoinDto{
         @NotNull
-        Integer score;
+        Float score;
         @Size(min = 5, max = 50)
         String contents;
